chore: update and fix kit
