+++ .gitignore
@@ -6,3 +6,5 @@ node_modules
 .env
 .env.*
 !.env.example
+.vercel
+.output
