+++ package.json
@@ -2,11 +2,11 @@
 	"name": "drawing-game",
 	"version": "3.0.0",
 	"scripts": {
-		"dev": "svelte-kit dev",
-		"build": "svelte-kit build",
+		"dev": "vite dev",
+		"build": "vite build",
 		"start": "node ./server",
 		"package": "svelte-kit package",
-		"preview": "svelte-kit preview",
+		"preview": "vite preview",
 		"prepare": "svelte-kit sync",
 		"lint": "prettier --check --plugin-search-dir=. . && eslint .",
 		"format": "prettier --write --plugin-search-dir=. ."
@@ -14,14 +14,16 @@
 	"devDependencies": {
 		"@sveltejs/adapter-node": "1.0.0-next.78",
 		"@sveltejs/kit": "next",
-		"eslint": "^8.18.0",
+		"eslint": "^8.19.0",
 		"eslint-config-prettier": "^8.5.0",
 		"eslint-plugin-svelte3": "^4.0.0",
 		"prettier": "^2.7.1",
 		"prettier-plugin-svelte": "^2.7.0",
 		"sass": "^1.53.0",
 		"svelte": "^3.48.0",
-		"svelte-preprocess": "^4.10.7"
+		"svelte-preprocess": "^4.10.7",
+		"vite-plugin-svelte-console-remover": "^1.0.10",
+		"vite": "^2.9.13"
 	},
 	"type": "module",
 	"dependencies": {
